# backend/api/swagger.yaml
---
swagger: "2.0"
info:
  description: "This is a server."
  version: "1.0.0"
  title: "Swagger_GestureWeb"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/api"
tags:
- name: "user"
  description: "Operations about user"
- name: "experiment"
  description: "Everything about your experiments"
schemes:
- "https"
- "http"
paths:
  /login:
    post:
      tags:
      - "user"
      summary: "Request authentication"
      description: "This method logs the user in."
      operationId: "login"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "User object that needs to be logged in."
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        405:
          description: "Invalid input"
      x-swagger-router-controller: "User"
  /experiment:
    get:
      tags:
      - "user"
      - "experiment"
      summary: "List the current user's experiments"
      description: "This method retrieves the whole list of experiments created by\
        \ the user that is logged in."
      operationId: "listExperiments"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "User object whose experiments are to be listed."
        required: true
        schema:
          $ref: "#/definitions/Experiment"
      responses:
        405:
          description: "Invalid user"
      x-swagger-router-controller: "User"
    post:
      tags:
      - "user"
      - "experiment"
      summary: "Create an experiment"
      description: "This method creates an experiment with the given parameters."
      operationId: "createExperiment"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Experiment object of the experiment that needs to be created."
        required: true
        schema:
          $ref: "#/definitions/Experiment"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Experiment"
        405:
          description: "Invalid input"
      x-swagger-router-controller: "User"
  /experiment/{experimentId}:
    get:
      tags:
      - "user"
      summary: "Request authentication"
      description: "This method logs the user in."
      operationId: "getExperiments"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "experimentId"
        in: "path"
        required: true
        type: "integer"
      responses:
        405:
          description: "Invalid input"
      x-swagger-router-controller: "User"
    put:
      tags:
      - "experiment"
      - "user"
      summary: "Update experiment"
      description: "This method updates an experiment given the user id."
      operationId: "updateExperiment"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "experimentId"
        in: "path"
        description: "ID of experiment to update"
        required: true
        type: "integer"
        format: "int64"
      responses:
        405:
          description: "Invalid input"
      x-swagger-router-controller: "Experiment"
    delete:
      tags:
      - "user"
      - "experiment"
      summary: "Delete experiment"
      description: "This method deletes the experiment."
      operationId: "deleteExperiment"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "experimentId"
        in: "path"
        required: true
        type: "integer"
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid experiment"
        404:
          description: "Experiment not found"
      x-swagger-router-controller: "User"
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      username:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
    example:
      firstName: "firstName"
      lastName: "lastName"
      password: "password"
      id: 0
      email: "email"
      username: "username"
  Experiment:
    type: "object"
    required:
    - "id"
    - "name"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      startDateTime:
        type: "string"
        format: "date-time"
      endDateTime:
        type: "string"
        format: "date-time"
    example:
      startDateTime: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      id: 0
      endDateTime: "2000-01-23T04:56:07.000+00:00"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"